requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
dj-database-url==2.3.0
    # via -r requirements.in
django==5.2
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==4.7.0
    # via -r requirements.in
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.0
    # via -r requirements.in
lxml==5.4.0
    # via
    #   -r requirements.in
    #   tools
pillow==11.2.1
    # via -r requirements.in
pyjwt==2.9.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
python-decouple==3.8
    # via -r requirements.in
pytils==0.4.3
    # via
    #   -r requirements.in
    #   tools
six==1.17.0
    # via
    #   -r requirements.in
    #   tools
sqlparse==0.5.3
    # via
    #   -r requirements.in
    #   django
tools==0.1.9
    # via -r requirements.in
typing-extensions==4.13.2
    # via
    #   -r requirements.in
    #   dj-database-url
stripe==7.0.0
